+++ docs/conf.py
@@ -369,7 +369,7 @@ texinfo_documents = [
 linkcode_resolve = make_linkcode_resolve(
     'xcp_d',
     (
-        'https://github.com/pennlinc/xcp_d/blob/' '{revision}/{package}/{path}#L{lineno}'  # noqa: FS003
+        'https://github.com/pennlinc/xcp_d/blob/{revision}/{package}/{path}#L{lineno}'  # noqa: FS003
     ),
 )