+++ xcp_d/cli/parser_utils.py
@@ -61,11 +61,14 @@ def _float_or_auto_or_none(string, is_parser=True):
 
     if floatarg < 0:
         raise error('Float argument must be nonnegative.')
-    return floatarg
+    return str(floatarg)
 
 
 def _restricted_float(x):
     """From https://stackoverflow.com/a/12117065/2589328."""
+    if x == 'auto':
+        return x
+
     try:
         x = float(x)
     except ValueError as exc: