+++ xcp_d/tests/test_cli_run.py
@@ -45,10 +45,12 @@ def base_opts():
         'motion_filter_order': None,
         'process_surfaces': 'auto',
         'atlases': ['Glasser'],
-        'dcan_correlation_lengths': None,
+        'min_coverage': 'auto',
+        'correlation_lengths': None,
         'despike': 'auto',
         'abcc_qc': 'auto',
         'linc_qc': 'auto',
+        'smoothing': 'auto',
         'combine_runs': 'auto',
         'output_type': 'auto',
         'fs_license_file': None,
@@ -276,9 +278,12 @@ def test_validate_parameters_linc_mode(base_opts, base_parser, capsys):
     assert opts.abcc_qc is False
     assert opts.linc_qc is True
     assert opts.file_format == 'cifti'
+    assert opts.min_coverage == 0.5
+    assert opts.smoothing == 6.0
+    assert opts.correlation_lengths == ['all']
 
     # --create-matrices is not supported
-    opts.dcan_correlation_lengths = [300]
+    opts.correlation_lengths = ['300']
     with pytest.raises(SystemExit, match='2'):
         parser._validate_parameters(deepcopy(opts), build_log, parser=base_parser)
 
@@ -298,19 +303,19 @@ def test_validate_parameters_abcd_mode(base_opts, base_parser, capsys):
 
     assert opts.abcc_qc is True
     assert opts.combine_runs is True
-    assert opts.dcan_correlation_lengths == []
+    assert opts.correlation_lengths == []
     assert opts.despike is True
     assert opts.fd_thresh == 0.3
     assert opts.file_format == 'cifti'
     assert opts.input_type == 'fmriprep'
     assert opts.linc_qc is True
-    assert opts.output_correlations is False
+    assert opts.min_coverage == 0.5
     assert opts.process_surfaces is True
+    assert opts.smoothing == 6.0
 
-    opts.dcan_correlation_lengths = ['300', 'all']
+    opts.correlation_lengths = ['300', 'all']
     opts = parser._validate_parameters(deepcopy(opts), build_log, parser=base_parser)
-    assert opts.dcan_correlation_lengths == ['300']
-    assert opts.output_correlations is True
+    assert opts.correlation_lengths == ['300', 'all']
 
     # --motion-filter-type is required
     opts.motion_filter_type = None
@@ -333,19 +338,19 @@ def test_validate_parameters_hbcd_mode(base_opts, base_parser, capsys):
 
     assert opts.abcc_qc is True
     assert opts.combine_runs is True
-    assert opts.dcan_correlation_lengths == []
+    assert opts.correlation_lengths == []
     assert opts.despike is True
     assert opts.fd_thresh == 0.3
     assert opts.file_format == 'cifti'
     assert opts.input_type == 'nibabies'
     assert opts.linc_qc is True
-    assert opts.output_correlations is False
+    assert opts.min_coverage == 0.5
     assert opts.process_surfaces is True
+    assert opts.smoothing == 6.0
 
-    opts.dcan_correlation_lengths = ['300', 'all']
+    opts.correlation_lengths = ['300', 'all']
     opts = parser._validate_parameters(deepcopy(opts), build_log, parser=base_parser)
-    assert opts.dcan_correlation_lengths == ['300']
-    assert opts.output_correlations is True
+    assert opts.correlation_lengths == ['300', 'all']
 
     # --motion-filter-type is required
     opts.motion_filter_type = None
@@ -354,6 +359,21 @@ def test_validate_parameters_hbcd_mode(base_opts, base_parser, capsys):
 
     stderr = capsys.readouterr().err
     assert "'--motion-filter-type' is required for" in stderr
+
+
+def test_validate_parameters_nichart_mode(base_opts, base_parser, capsys):
+    """Test parser._validate_parameters with nichart mode."""
+    opts = deepcopy(base_opts)
+    opts.mode = 'nichart'
+
+    # linc mode doesn't use abcc_qc but does use linc_qc
+    opts = parser._validate_parameters(deepcopy(opts), build_log, parser=base_parser)
+
+    assert opts.abcc_qc is False
+    assert opts.linc_qc is True
+    assert opts.file_format == 'nifti'
+    assert opts.min_coverage == 0.4
+    assert opts.smoothing == 0
 
 
 def test_validate_parameters_none_mode(base_opts, base_parser, capsys):
@@ -372,9 +392,11 @@ def test_validate_parameters_none_mode(base_opts, base_parser, capsys):
     assert "'--file-format' is required for 'none' mode." in stderr
     assert "'--input-type' is required for 'none' mode." in stderr
     assert "'--linc-qc' (y or n) is required for 'none' mode." in stderr
+    assert "'--min-coverage' is required for 'none' mode." in stderr
     assert "'--motion-filter-type' is required for 'none' mode." in stderr
     assert "'--nuisance-regressors' is required for 'none' mode." in stderr
     assert "'--output-type' is required for 'none' mode." in stderr
+    assert "'--smoothing' is required for 'none' mode." in stderr
     assert "'--warp-surfaces-native2std' (y or n) is required for 'none' mode." in stderr
 
     opts.abcc_qc = False
@@ -385,10 +407,12 @@ def test_validate_parameters_none_mode(base_opts, base_parser, capsys):
     opts.file_format = 'nifti'
     opts.input_type = 'fmriprep'
     opts.linc_qc = False
+    opts.min_coverage = 0.5
     opts.motion_filter_type = 'none'
     opts.output_type = 'censored'
     opts.params = '36P'
     opts.process_surfaces = False
+    opts.smoothing = 0
     opts = parser._validate_parameters(deepcopy(opts), build_log, parser=base_parser)
 
 
@@ -397,7 +421,7 @@ def test_validate_parameters_other_mode(base_opts, base_parser, capsys):
     opts = deepcopy(base_opts)
     opts.mode = 'other'
 
-    with pytest.raises(AssertionError, match="Unsupported mode 'other'"):
+    with pytest.raises(AssertionError, match='Unsupported mode "other"'):
         parser._validate_parameters(deepcopy(opts), build_log, parser=base_parser)
 
 
@@ -435,7 +459,7 @@ def test_build_parser_01(tmp_path_factory):
     opts = parser_obj.parse_args(args=test_args, namespace=None)
     assert opts.fmri_dir == data_path
     assert opts.output_dir == out_path
-    assert opts.dcan_correlation_lengths == ['all', 300, 480]
+    assert opts.correlation_lengths == ['all', '300.0', '480.0']
 
 
 def test_build_parser_02(tmp_path_factory):
@@ -472,7 +496,7 @@ def test_build_parser_02(tmp_path_factory):
     opts = parser_obj.parse_args(args=test_args, namespace=None)
     assert opts.fmri_dir == data_path
     assert opts.output_dir == out_path
-    assert opts.dcan_correlation_lengths == ['all', 300, 480]
+    assert opts.correlation_lengths == ['all', '300.0', '480.0']
 
 
 @pytest.mark.parametrize(