+++ xcp_d/utils/modified_data.py
@@ -275,6 +275,7 @@ def calculate_exact_scans(exact_times, scan_length, t_r, bold_file):
     float_times = []
     non_float_times = []
 
+    exact_times = [time for time in exact_times if time != 'all']
     for time in exact_times:
         try:
             float_times.append(float(time))
@@ -293,8 +294,7 @@ def calculate_exact_scans(exact_times, scan_length, t_r, bold_file):
 
     if non_float_times:
         LOGGER.warning(
-            f'Non-float values {non_float_times} in {os.path.basename(bold_file)} '
-            'will be ignored.'
+            f'Non-float values {non_float_times} in {os.path.basename(bold_file)} will be ignored.'
         )
 
     exact_scans = [int(t // t_r) for t in retained_exact_times]